# lib/api-zod/src/generated/types/appointmentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rural Telemedicine Portal API
 * OpenAPI spec version: 0.1.0
 */

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
  cancelled: "cancelled",
} as const;