# lib/api-zod/src/generated/types/appointmentStatusUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rural Telemedicine Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from "./appointmentStatus";

export interface AppointmentStatusUpdate {
  status: AppointmentStatus;
}