# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rural Telemedicine Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: string;
  name: string;
  specialty: string;
}