# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rural Telemedicine Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: string;
  patientId: string;
  patientName: string;
  name: string;
  mimeType: string;
  sizeBytes: number;
  dataUrl: string;
  note?: string | null;
  createdAt: Date;
}