# lib/api-zod/src/generated/types/newDocumentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rural Telemedicine Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface NewDocumentInput {
  name: string;
  mimeType: string;
  /** Data URL (data:<mime>;base64,...) */
  dataUrl: string;
  note?: string | null;
}