# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rural Telemedicine Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Language } from "./language";
import type { Role } from "./role";

export interface Profile {
  id: string;
  role: Role;
  name: string;
  language: Language;
  specialty?: string | null;
  village?: string | null;
}