# lib/api-zod/src/generated/types/symptomCondition.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rural Telemedicine Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SymptomCondition {
  name: string;
  /** 0-100 */
  confidence: number;
  advice: string;
}